refactor(@dpc-sdp/ripple-tide-search): consolidated common results values into composable
